% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSettings.R
\name{addSettings}
\alias{addSettings}
\title{Add Settings to list param}
\usage{
addSettings(script, tipo, configuracao = NULL)
}
\arguments{
\item{script}{Script is a string that contains the script}

\item{tipo}{Type is one of the possibilities of parameter of evaluationVolumeAdvanced like: "models" or "graphs" or "statistics" or "save"}

\item{configuracao}{Configuration is a list that contains the settings for evaluation in the function EvaluateSpecial ()}
}
\value{
list of settings
}
\description{
This function assigns if the parameter does not exist or adds the parameter to the configuration. The variable assigned to the parameter is present in the last line of the script
}
