% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliaModeloEspecial.R
\name{avaliaModeloEspecial}
\alias{avaliaModeloEspecial}
\title{Avalia Model Special}
\usage{
avaliaModeloEspecial(configuracao)
}
\arguments{
\item{configuracao}{It is the list of settings}
}
\value{
result of avaliaVolumeAvancado2
}
\description{
This function is a wizard to use the functionValueValueAdvanced2 configuration: a list of configurations that will be passed as parameter to function avaliaVolumeAvancado2
}
