% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criaModeloExclusivo2.R
\name{criaModeloExclusivo2}
\alias{criaModeloExclusivo2}
\title{Create Exclusive Model for a database}
\usage{
criaModeloExclusivo2(modeloGenerico, variaveis, palpite = NULL)
}
\arguments{
\item{modeloGenerico}{model of pattern criaModeloGenerico}

\item{variaveis}{list of name fields (strings) in database and model, the order of variables matter}

\item{palpite}{string containing start values of function of regression}
}
\value{
will be returned a function with exclusive model
}
\description{
this function returns a unique model is variable receive each mapeda variable ex .: criaModeloExclusivo (modeloCamposLeite, c ("age1", "age2", "bai1", "s"))
}
