% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstsMeasurements.R
\name{firstsMeasurements}
\alias{firstsMeasurements}
\title{First Measurements per Group}
\usage{
firstsMeasurements(dtFrame, group = "parcela", age = "idadearred")
}
\arguments{
\item{dtFrame}{It is the database that contains the data}

\item{group}{is the field that represents the groups default parcela}

\item{age}{default idadearred is the name of field containing the values of the group}
}
\value{
data.frame dtFrame with the field "primeiraMedicao"
}
\description{
this function generates a field "primeiraMedicao" that contains the lowest age for each item in the group both derived from dtFrame
}
\examples{
dtf = data.frame(
grupo = c(1,2,3,4,1,2,3,4),
medicoes= c(10,20,30,40,5,30,1,52))
firstsMeasurements(dtf, group = "grupo", age = "medicoes")
}
