% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllEstatisticsFn.R
\name{getAllEstatisticsFn}
\alias{getAllEstatisticsFn}
\title{Get All Functions of Statistics of ITGM}
\usage{
getAllEstatisticsFn()
}
\value{
collection of functions statistics
}
\description{
this function allows you to obtain all functions available in ITGM for model evaluation.
}
