% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllGraphicsFn.R
\name{getAllGraphicsFn}
\alias{getAllGraphicsFn}
\title{Get All Functions of Graphics of ITGM}
\usage{
getAllGraphicsFn()
}
\value{
collection of functions graphics
}
\description{
this function allows you to obtain all functions available in ITGM for model evaluation.
}
