% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraphicVolumeTotal.R
\name{getGraphicVolumeTotal}
\alias{getGraphicVolumeTotal}
\title{Get rbokeh grapic observed versus estimated}
\usage{
getGraphicVolumeTotal(
  titulo = "observadoXestimado",
  nome = "observadoXestimado",
  observado,
  estimado,
  showTestF = TRUE,
  save = NULL,
  labsX = "observado",
  labsy = "estimado",
  vetorial = T,
  desc = NULL,
  data = NULL,
  mapeamento = list(idade2 = "idadearred2", parcela = "parcela", areacorr = "areacorr"),
  ...
)
}
\arguments{
\item{titulo}{is the title graphic}

\item{nome}{name of file case save}

\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{showTestF}{draw results of test F in graphic?}

\item{save}{If you want to save enter the directory as a string}

\item{labsX}{label x}

\item{labsy}{label y}

\item{vetorial}{save picture in vector type? (Default TRUE)}

\item{desc}{description of plot geted from avaliavolumeavancado}

\item{data}{database to retireve fields to id}

\item{mapeamento}{name of fields on base}

\item{...}{optionals params to plot graphic in ITGM}
}
\value{
will be returned a rbokeh graphic
}
\description{
this function displays/saves/returns a Graphical rbokeh illustrating the difference between the observed and estimated
}
