% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelosLiteraturaExclusivos.R
\name{getModelosLiteraturaExclusivos}
\alias{getModelosLiteraturaExclusivos}
\title{get Models Literature Exclusives}
\usage{
getModelosLiteraturaExclusivos(mapeamento = c("idade1", "idade2", "bai1", "s"))
}
\arguments{
\item{mapeamento}{list of names fields of database will work "idade1", "idade2", "bai1", "s"}
}
\value{
will be returned list of function with exclusive model
}
\description{
This function creates and returns an array with the usual models mapped to the mapping vector
}
