% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelosLiteraturaGenericos.R
\name{getModelosLiteraturaGenericos}
\alias{getModelosLiteraturaGenericos}
\title{get Generic Model Literature}
\usage{
getModelosLiteraturaGenericos()
}
\value{
will be returned list of models generic
}
\description{
This function creates and returns an array with the usual models of literature
}

