% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelsExclusiveOfLiterature.R
\name{getModelsExclusiveOfLiterature}
\alias{getModelsExclusiveOfLiterature}
\title{get Models Literature Exclusives}
\usage{
getModelsExclusiveOfLiterature(mapper = c("idade1", "idade2", "bai1",
  "s"))
}
\arguments{
\item{mapper}{list of names fields of database will work "idade1", "idade2", "bai1", "s"}
}
\value{
will be returned list of function with exclusive model
}
\description{
This function creates and returns an array with the usual models mapped to the mapping vector
}
