% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelsGenericOfLiterature.R
\name{getModelsGenericOfLiterature}
\alias{getModelsGenericOfLiterature}
\title{get Generic Model Literature}
\usage{
getModelsGenericOfLiterature()
}
\value{
will be returned list of models generic
}
\description{
This function creates and returns an array with the usual models of literature
}
