% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRB.R
\name{plotRB}
\alias{plotRB}
\title{Get rbokeh grapic observed versus estimated}
\usage{
plotRB(
  titulo = "observadoXestimado",
  nome = "observadoXestimado",
  observado,
  estimado,
  id = NULL,
  save = NULL,
  labsX = "observado",
  labsy = "estimado",
  showTestF = TRUE,
  colID = "idLabel",
  sobnome = "ObservadoXEstimado",
  data = NULL,
  diretorio = NULL,
  datadiretorio = NULL,
  HTTPS = FALSE,
  totalizar = NULL,
  campoID = NULL,
  mapeamento = list(idade2 = "idadearred2", parcela = "parcela", areacorr = "areacorr"),
  ...
)
}
\arguments{
\item{titulo}{is the title graphic}

\item{nome}{name of file case save}

\item{observado}{list containing the observations of variable}

\item{estimado}{list containing estimates of variable}

\item{id}{list containing 'id' of individuals}

\item{save}{If you want to save enter the directory as a string}

\item{labsX}{label x}

\item{labsy}{label y}

\item{showTestF}{draw results of test F in graphic?}

\item{colID}{label of ids}

\item{sobnome}{optional name to append of file name graphic}

\item{data}{database to retireve fields to id}

\item{diretorio}{database to retireve fields to id}

\item{datadiretorio}{optional abstract path of directory}

\item{HTTPS}{script in https?}

\item{totalizar}{generat graphic of volume total}

\item{campoID}{name of parcela field in data}

\item{mapeamento}{name of fields on base}

\item{...}{optionals params to plot graphic in ITGM}
}
\value{
will be returned a rbokeh graphic
}
\description{
this function displays/saves/returns a Graphical rbokeh illustrating the difference between the observed and estimated
}
