% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparaFormula.R
\name{preparaFormula}
\alias{preparaFormula}
\title{Prepare formula}
\usage{
preparaFormula(formula)
}
\arguments{
\item{formula}{It is the string formula to use in modelo}
}
\value{
formula handled
}
\description{
this function prepare formula to use in models
}
