% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project Volume based in Ages}
\usage{
project(firstAge = NaN, lastAge, fitDAP, fitHT, base, mapper = list(age1 =
  "idade1", age2 = "idade2", dap1 = "dap1", dap2 = "dap2", ht1 = "ht1", ht2 =
  "ht2"), calcVolume = calculaVolumeDefault, withoutBaseFields = F)
}
\arguments{
\item{firstAge}{early age. if only one age use NaN}

\item{lastAge}{late age for project or the age at which one wants to get the volume}

\item{fitDAP}{an adjustment of the return type of a function lm() from dap}

\item{fitHT}{an adjustment of the return type of a function lm() from ht}

\item{base}{a dataset to project}

\item{mapper}{the mapping for the name of the old fields age, dap and ht in base}

\item{calcVolume}{function to calc volume based base, dap e ht,  default calculaVolumeDefault of Fgmutils}

\item{withoutBaseFields}{want returned projected volume no contains the fields of the base? default no}
}
\description{
this function provides a list of volume projections in a future age or from one to another future age
}

