% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeColsSuspicious.R
\name{removeColsSuspicious}
\alias{removeColsSuspicious}
\title{Create function with generic model}
\usage{
removeColsSuspicious(base, colsRm)
}
\arguments{
\item{base}{is the data with fields to remove}

\item{colsRm}{is the collections of name fields suspicius to remove of base}
}
\value{
will be returned function with generic model to map to a base
}
\description{
This function creates a generic model that will be a funcao that has parameters for the variables that can be mapped to each different base. her return will be a generic model that should be mapped to be used by the function avaliaEstimativas
}
